refactor(logger): inline structuredLogger into CLI, trim LogUtilsOptions
Step 3 — `src/logger/structuredLogger.ts` had a single consumer (the
CLI's `runtimeHost.ts` / `logSinks.ts`) after the trace channel took
over group context. Inlined the ~30 lines into
`packages/cli/src/runtime/logSinks.ts` as `createCliLogger`, deleted
the original module + its now-redundant test. `LogRecord` lost the
always-empty `groups` field, simplifying `StderrTextSink`.

Step 4 — `LogUtilsOptions` had four fields; three (`groupId`,
`messageType`, `isAgent`) had zero functional callers across ~70
`logUtils.{level}()` consumers. Trimmed the type to `{ data?: unknown }`
and hard-coded `isAgent: false` in `logAt` (only the trace-subscriber
path needs the per-stream channel, and that path goes through
`attachChannelSubscriber` → `writeLine` directly, bypassing `logAt`).
`MessageType` import dropped along with it.

Net: −78 lines. `@logger/` is now five focused modules instead of six.

Verified: typecheck + test-kernel + lint all clean; 1042/1042 tests pass.
